+++ contracts/scripts/bridge/utils.ts
@@ -1,0 +1,96 @@
+const l2Provider = new ethers.providers.JsonRpcProvider(
+  "http://localhost:4011"
+);
+
+const l1Provider = new ethers.providers.JsonRpcProvider(
+  "http://localhost:8545"
+);
+
+const l1BridgeAddr = "0xE7C2a73131dd48D8AC46dCD7Ab80C8cbeE5b410A";
+const l2BridgeAddr = "0xF6168876932289D073567f347121A267095f3DD6";
+const l1OracleAddress = "0x2E983A1Ba5e8b38AAAeC4B440B9dDcFBf72E15d1";
+const rollupAddress = "0xF6168876932289D073567f347121A267095f3DD6";
+
+export async function getSignersAndContracts() {
+  const l1Bridger = new ethers.Wallet(
+    process.env.BRIDGER_PRIVATE_KEY,
+    l1Provider
+  );
+  const l2Bridger = new ethers.Wallet(
+    process.env.BRIDGER_PRIVATE_KEY,
+    l2Provider
+  );
+
+  const l1Relayer = new ethers.Wallet(
+    process.env.SEQUENCER_PRIVATE_KEY,
+    l1Provider
+  );
+  const l2Relayer = new ethers.Wallet(
+    process.env.SEQUENCER_PRIVATE_KEY,
+    l2Provider
+  );
+
+  const L1StandardBridgeFactory = await ethers.getContractFactory(
+    "L1StandardBridge",
+    l1Bridger
+  );
+  const l1StandardBridge = L1StandardBridgeFactory.attach(l1BridgeAddr);
+
+  const L2StandardBridgeFactory = await ethers.getContractFactory(
+    "L2StandardBridge",
+    l2Bridger
+  );
+  const l2StandardBridge = L2StandardBridgeFactory.attach(l2BridgeAddr);
+
+  const l1PortalAddress = await l1StandardBridge.PORTAL_ADDRESS();
+  const L1PortalFactory = await ethers.getContractFactory(
+    "L1Portal",
+    l1Bridger
+  );
+  const l1Portal = L1PortalFactory.attach(l1PortalAddress);
+
+  const l2PortalAddress = await l2StandardBridge.PORTAL_ADDRESS();
+  const L2PortalFactory = await ethers.getContractFactory(
+    "L2Portal",
+    l2Relayer
+  );
+  const l2Portal = L2PortalFactory.attach(l2PortalAddress);
+
+  const L1OracleFactory = await ethers.getContractFactory(
+    "L1Oracle",
+    l2Relayer
+  );
+  const l1Oracle = L1OracleFactory.attach(l1OracleAddress);
+
+  const RollupFactory = await ethers.getContractFactory("Rollup", l1Relayer);
+  const rollup = await RollupFactory.attach(rollupAddress);
+
+  l1Portal.on("*", (...args) => console.log({ ...args }));
+  l2Portal.on("*", (...args) => console.log({ ...args }));
+  l1StandardBridge.on("*", (...args) => console.log({ ...args }));
+  l2StandardBridge.on("*", (...args) => console.log({ ...args }));
+
+  return {
+    l1Provider,
+    l2Provider,
+    l1Bridger,
+    l2Bridger,
+    l1Relayer,
+    l2Relayer,
+    l1Portal,
+    l2Portal,
+    l1StandardBridge,
+    l2StandardBridge,
+    l1Oracle,
+    rollup,
+  };
+}
+
+export function getStorageKey(messageHash: string) {
+  return ethers.utils.keccak256(
+    ethers.utils.defaultAbiCoder.encode(
+      ["bytes32", "uint256"],
+      [messageHash, 0]
+    )
+  );
+}
